Which language the comment comment is about
Any time you're describing a language for describing a language, it can get confusing.  I found this line in the `README` confusing, because I didn't know whether it meant "You can add a nearley file, `#` starts a comment" or "This grammar defines a language that ignores comments beginning with `#`.  There's no use of `#` in the example that clarifies it.  Based on `http://nearley.js.org/www/railroad-demo.html`, I'm sure you meant the former, so how about describe it with a second example?