[WIP] Improve the docs (#254)
* Add some lexer stuff to README

* readme: remove trailing spaces

* readme: change the description to "Simple parsing in JavaScript"

nearley can run in browsers too.

* readme: update the table of contents

Run `doctoc`. The ToC got outdated.

* readme: make Markdown style consistent

Use consistent style: `#` headers, no extra newlines, `-` lists.
I hope this is the rigth style, this is how the majority of the file looked like.

* readme: add an intro and usage guide for beginners

* readme: reorder and reword parser specification

* readme: make all example code more readable, update to ES6

* docs: move `glossary.md` and `how-to-grammar-good.md` to `docs/`

* docs: move making REPL and accessing parse table into their own files

These two topics are covered in the section "Using a parser", which doesn't make much sense.
The basic usage is already at the top of readme, so let's remove "Using a parser" completely.

* docs: add a clear Moo example to readme, separate out custom lexers

* docs: add a link in readme to "How to grammar good"

* docs: add a complete example of usage in browsers

* docs: use `require()` instead of `import`

* readme: link to the indentation-aware lexer by nathan

* docs: add an example of generating CST and AST

* readme: move recipies to section "Recipies"

* Make clear using `nearleyc` in a browser is unusual.

this is a test commit. Can I commit to a PR? That seems odd.

* Fix a bunch of my review comments

:-)

* Add doctoc script to package.json